add new system for placing objects on landscape/other objects
This commits adds new module called ObjectPlacementModule, which from
now on will manage placement objects or another objects.

This module adds action "Place on landscape" and "Snap to landscape"
to the main toolbar, and to the Edit menu.

Existent actions that were defined in .ui file, were removed.

With the new module comes new system ObjectPlacementSystem, which now
manages object placement on landscape and other objects.

We also override and/remove entity methods, since we want to know, if
the landscape currently present on the scene, and to store it in the
system, since we don't want to find landscape every time we need it
for something.

However, since current Place on landscape/Snap to landscape
functionality is tightly coupled with EntityModificationSystem, method
moving from it to the new system will be done later. For now we just
have pointer to EntityModificationSystem in the ObjectPlacementSystem,
and proxy calls to it. This will be changed in the future. Hopefully.

Also all connection methods and slots from mainwindow.cpp and
ModificationWidget.cpp were removed, as our controls became TArc-based
and therefore reflective.

DF-14792


Former-commit-id: a8a5648a8725e0fdd4989b5bbebf09e2b4c98836