fix ray tracing not returning collision even if it was found
PolygonGroup doesn't have OctTrees built by default, it's built on
demand, specifically, on first query.

But in RayTrace method we accessed octree field bypassing the getter,
so if it haven't been generated earlier, we got nullptr.

If we change direct access to access via getter, it will generate
OctTree and everything will be ok.

DF-14792


Former-commit-id: 5476fac05cd288206a8329e0bbda23cd9fd51bbe