fix HashToString returning string with extra \0
New HashToString created string from buffer of type char* and its
size. However, buffer already included \0, so after string
have been created, it had and extra \0 at the end, and when we tried
to append some other string to it, that extra zero byte prevented the
append.

To fix this, we remove buffer size from string construction, which
fixes additional \0 at the end of the string.

in scope of DF-14822


Former-commit-id: 7888496ea5b6a768f2a554e33439e2659aaca93d