switch CameraSystem to new saving mechanism (PrepareToSave method)
As part of decoupling process we teach SceneCameraSystem how to
prepare for scene saving in generic manner -- by generating
DavaCommand with Undo and Redo methods overriden, where Redo makes all
necessary changes and Undo undoes them back.

In case of SceneCameraSystem, we need to remove debug cameras from the
scene before saving, store it for some time and then bring it back to
the scene after save have been finished.

We also enable camera to store debug camera entity, so we can access
it fast insted of searching all the camera entities for "editor."
prefix.

in scope if DF-14822


Former-commit-id: 141b3e8ad12cf5d4c1c91417ca60239354f1268a