add local Project-unique scene properties to Resource Editor
Here we enable ResouceEditor to have per-scene properties, which
automatically created on scene creation, and are local to scene inside
the project.

We want to store settings that are linked to the particular user,
particular project and particular scene inside this project, in that
exact user. That is, that information is not stored in the project,
but insted in user documents directory (~doc in terms of
DAVA::FileSystem).

Newly created scenes must also have temporary storages, which are
discarded on each new scene creation, but might be used as a source
for copying properties file when saving the scene.

On "Save As" action storage from prrevious scene must be copied to new
location.

On scene loading we load corresponding properties file, or if it can't
be found we create a new one.

Properties for projects are located in
~doc/SceneProperties/<PROJECT_HASH>/, where <PROJECT_HASH> is and md5
sum of an absolute path to project directory.

Project for each scene settings reside in
~/doc/SceneProperties/<PROJECT_HASH>/path/to/scene/scene_name.sc2,
i.e. in ~/doc/SceneProperties/<PROJECT_HASH> directory followed by
relative scene path in actual project, and name is the same as
scene's.
Default scene properties files lie in ~/doc/SceneProperties/Default.

To support such logic, we need to make following changes:

Enable PropertiesHolder to copy its data storage to new location and
re-initialize itself by new path -- so we can copy temporary scene
property files when we save them, or when we save scene under
different name.

Add PropertiesHolder to the SceneData as unique_ptr, and create
function to create it, since between creation of SceneData object and
moment when scene path (or lack of thereof) is determined, can pass
quite a period of time.

We don't care if we need to create new PropertiesHolder or copy old
one to new place, so we abstract this in SceneData, because we have to
copy property storage only when saving scene under a new name.

Other components, which need to save scene-local properties should get
SceneData from current context and call GetPropertiesSection method,
which returns PropertiesItem, which they can use at their pleasure.

DF-14822


Former-commit-id: 980594c282c59d08a550c07c46603764bbe2aefa