fix ObjectPlacementSystem placing object to itself
When we cast a ray towards the earth, it didn't know anything about
objects it was originated from. So if that object was aligned so its
pivot point is up, and most of geometry down, ray found intersection
with that object, causing object to be aligned as if it was places on
itself.

This behaviour is unwanted, so we fix it by collecting all selected
entities' render objects and passing it to RayTrace method as
ignoreObjects parameter. So if algorithm found collisions with object
from that vector, this collision is discarded and ray tracing proceeds
until it finds next object.

DF-14792
DF-15133


Former-commit-id: 4e34b15924b40880ab6d509d8dce3e466326b063