enable SceneCameraSystem to save active camera between RE launches
To make this possible we get current active camera entity name and
save it to the scene properties. And on loading, we traverse all scene
cameras and look for camera with saved name. If we don't find such a
camera, we do nothig. Otherwise, set it as current scene camera.

DF-12706


Former-commit-id: 256188e8da778e2f8a75afd5b35decfa5f0b4b82