fix RE landscape tile editor coordinates alignment to integers
For historical reasons, landscape pointer (that blue circle) position
coordinates were aligned to nearest integers. Due to this, when
landscape size was small (like 100), and brush (and pointer) size was
also small (like 2), it aligned to integer grid, which is not right
and greatly annoyed artists.

Here we remove this coordinate flooring thus fixing the problem with
that unintended grid snap.

DF-15314


Former-commit-id: 238a745829e72b819ef78dccf880f2af29caa0b4