fix SO building too long when lanscape has Height Field component
When Height Field component added to the landscape entity, debug
wireframe is enabled for it. However, wireframe's material is not
added to debug layer, and instead is added to opaque layer. That
caused it to be added to static occlusion calculations.

We fix this by providing correct layer for wireframe material.

DF-15075


Former-commit-id: 78fcddeeb7f0c3e77e61e041b8f577e3dca92820