fix RE crash when reloading entity model from Scene Tree
EntityChangeParentCommand usage for adding new entity is unclear and
not really a valid use case when reloading model. So we change that
command to EntityAddCommand.

However, we want to insert reloaded entity to exact place from which
we reloaded old one, so we also enhance EntityAddCommand with ability
to insert entity after other one, when specified.

DF-16141


Former-commit-id: 0905f7af8520134054e1d9867df95608c252472d